refactor(plugins): Move plugins-api to api package (#502)
