+++ kork-plugins/src/test/java/com/netflix/spinnaker/kork/plugins/testplugin/unsafe/UnsafeTestExtension.java
@@ -16,7 +16,7 @@
 
 package com.netflix.spinnaker.kork.plugins.testplugin.unsafe;
 
-import com.netflix.spinnaker.kork.plugins.SpinnakerExtension;
+import com.netflix.spinnaker.kork.plugins.api.SpinnakerExtension;
 import com.netflix.spinnaker.kork.plugins.testplugin.api.TestExtension;
 import org.pf4j.Extension;
